requirements.txt
streamlit
urllib3
Pillow
docx2txt
PyPDF2
python-dotenv
langchain